backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pillow==10.4.0
python-multipart==0.0.9

numpy==1.26.4
scikit-image==0.24.0
torchxrayvision==1.4.0
